new indexer with only one event tracking on both chains
