+++ envio-indexer/config.yaml
@@ -1,0 +1,19 @@
+name: Deposit indexer
+contracts:
+- name: SBCDepositContract
+  handler: src/EventHandlers.ts
+  events:
+  - event: DepositEvent(bytes pubkey, bytes withdrawal_credentials, bytes amount, bytes signature, bytes index)
+networks:
+- id: 100
+  start_block: 19469076
+  contracts:
+  - name: SBCDepositContract
+    address:
+    - 0x0B98057eA310F4d31F2a452B414647007d1645d9
+- id: 10200
+  start_block: 155434
+  contracts:
+  - name: SBCDepositContract
+    address:
+    - 0xb97036A26259B7147018913bD58a774cf91acf25
