--- envio/schema.graphql
@@ -1,60 +1,0 @@
-type ClaimRegistryUpgradeable_ClaimBatch {
-  id: ID!
-  caller: String!
-  withdrawalAddresses: [String!]!
-}
-
-type ClaimRegistryUpgradeable_Initialized {
-  id: ID!
-  version: BigInt!
-}
-
-type ClaimRegistryUpgradeable_OwnershipTransferred {
-  id: ID!
-  previousOwner: String!
-  newOwner: String!
-}
-
-type ClaimRegistryUpgradeable_Register {
-  id: ID!
-  user: String!
-}
-
-type ClaimRegistryUpgradeable_Unregister {
-  id: ID!
-  user: String!
-}
-
-type ClaimRegistryUpgradeable_UpdateConfig {
-  id: ID!
-  user: String!
-  oldTime: BigInt!
-  newTime: BigInt!
-  oldAmount: BigInt!
-  newAmount: BigInt!
-}
-
-type ClaimRegistryUpgradeable_Upgraded {
-  id: ID!
-  implementation: String!
-}
-
-type SBCDepositContract_DepositEvent {
-  id: ID!
-  pubkey: String!
-  withdrawal_credentials: String!
-  amount: String!
-  signature: String!
-  index: String!
-}
-
-type SBCDepositContract_Paused {
-  id: ID!
-  account: String!
-}
-
-type SBCDepositContract_Unpaused {
-  id: ID!
-  account: String!
-}
-
