+++ hooks/use-deposit.ts
@@ -6,10 +6,6 @@ import {
 import { ContractNetwork } from "@/utils/contracts";
 import ERC677ABI from "@/utils/abis/erc677";
 import { formatUnits, parseUnits } from "viem";
-// import { loadCachedDeposits } from "@/utils/deposit";
-import { getPublicClient } from "wagmi/actions";
-import { config } from "@/wagmi";
-// import { fetchDeposit } from "@/utils/fetchEvents";
 import useBalance from "./use-balance";
 import { gql, useApolloClient } from '@apollo/client';
 
@@ -26,13 +22,14 @@ type DepositDataJson = {
 };
 
 const GET_DEPOSIT_EVENTS = gql`
-  query MyQuery($pubkeys: [String!]) {
+  query MyQuery($pubkeys: [String!], $chainId: Int!) {
     SBCDepositContract_DepositEvent(
       limit: 10, 
       where: { 
         pubkey: { 
           _in: $pubkeys
-        }
+        },
+        chainId: {_eq: $chainId}
       }
     ) {
       id
@@ -52,9 +49,6 @@ function useDeposit(contractConfig: ContractNetwork | undefined, address: `0x${s
   const [isBatch, setIsBatch] = useState(false);
   const [filename, setFilename] = useState("");
   const { balance, refetchBalance } = useBalance(contractConfig, address);
-  const client = getPublicClient(config, {
-    chainId: chainId as 100 | 10200 | 31337,
-  });
   const isWrongNetwork = contractConfig === undefined;
   const { data: depositHash, writeContract } = useWriteContract();
   const { isSuccess: depositSuccess } = useWaitForTransactionReceipt({
@@ -105,28 +99,12 @@ function useDeposit(contractConfig: ContractNetwork | undefined, address: `0x${s
           );
         }
 
-        // const { deposits: existingDeposits, lastBlock: fromBlock } =
-        //   await loadCachedDeposits(
-        //     chainId === 31337 ? 10200 : chainId,
-        //     contractConfig.depositStartBlockNumber
-        //   );
-
-        // const events = await fetchDeposit(
-        //   contractConfig.addresses.deposit,
-        //   fromBlock,
-        //   client
-        // );
-
-        // let pks = events.map((e) => e.args.pubkey);
-        // pks = pks.concat(existingDeposits);
-        // console.log(pks);
-        // console.log(`Found ${pks.length} existing deposits`);
-
         const pksFromFile = deposits.map((d) => `0x${d.pubkey}`);
         const { data } = await apolloClient.query({
           query: GET_DEPOSIT_EVENTS,
           variables: {
             pubkeys: pksFromFile,
+            chainId: chainId,
           },
         });
         const existingDeposits = data.SBCDepositContract_DepositEvent.map((d: { pubkey: string }) => d.pubkey);
@@ -192,7 +170,7 @@ function useDeposit(contractConfig: ContractNetwork | undefined, address: `0x${s
 
       return { deposits: newDeposits, hasDuplicates, _isBatch };
     },
-    [address, contractConfig, balance]
+    [contractConfig, apolloClient, chainId]
   );
 
   const setDepositData = useCallback(
@@ -283,7 +261,7 @@ function useDeposit(contractConfig: ContractNetwork | undefined, address: `0x${s
         );
       }
     }
-  }, [address, deposits, isBatch, refetchBalance]);
+  }, [contractConfig, deposits, isBatch, refetchBalance, writeContract]);
 
   useEffect(() => {
     if (depositSuccess) {
